chore: TextEncoderStream_polyfill
