+++ packages/qwik-city/middleware/cloudflare-pages/index.ts
@@ -5,6 +5,7 @@ import type {
 import {
   mergeHeadersCookies,
   requestHandler,
+  TextEncoderStream_polyfill,
 } from '@builder.io/qwik-city/middleware/request-handler';
 import { getNotFound } from '@qwik-city-not-found-paths';
 import { isStaticPath } from '@qwik-city-static-paths';
@@ -20,7 +21,7 @@ export function createQwikCity(opts: QwikCityCloudflarePagesOptions) {
     // this will throw if CF compatibility_date < 2022-11-30
     new globalThis.TextEncoderStream();
   } catch (e) {
-    globalThis.TextEncoderStream = TextEncoderStream as any;
+    globalThis.TextEncoderStream = TextEncoderStream_polyfill as any;
   }
   const qwikSerializer = {
     _deserializeData,
@@ -137,40 +138,4 @@ export interface PlatformCloudflarePages {
   request: Request;
   env?: Record<string, any>;
   ctx: { waitUntil: (promise: Promise<any>) => void };
-}
-
-const resolved = Promise.resolve();
-
-class TextEncoderStream {
-  // minimal polyfill implementation of TextEncoderStream
-  // since Cloudflare Pages doesn't support readable.pipeTo()
-  _writer: any;
-  readable: any;
-  writable: any;
-
-  constructor() {
-    this._writer = null;
-    this.readable = {
-      pipeTo: (writableStream: any) => {
-        this._writer = writableStream.getWriter();
-      },
-    };
-    this.writable = {
-      getWriter: () => {
-        if (!this._writer) {
-          throw new Error('No writable stream');
-        }
-        const encoder = new TextEncoder();
-        return {
-          write: async (chunk: any) => {
-            if (chunk != null) {
-              await this._writer.write(encoder.encode(chunk));
-            }
-          },
-          close: () => this._writer.close(),
-          ready: resolved,
-        };
-      },
-    };
-  }
 }
